波形メモリ音色定義の文法をSiON互換にしました。 ある程度動作確認ができたためです。 このため sion-adapter.js を pre と postに分割しました。
