+++ README.md
@@ -43,12 +43,12 @@ FM2OP音色定義：
 　※開発中。仕様変更もありえます。
 
 波形メモリ音色定義：
-#WAVBn,X(64文字)
+#WAVBn{X(64文字)}
  n: 音色番号
- X: [0-9A-F]
-　※現在実装速度優先で大文字のみ対応しています。
+ X: [0-9A-Fa-f]
+　※現在実装速度優先でトラックごとに独立しています。
 　※#FM2OP@nと#WAVBnで同じ番号を指定した場合は後着優先です。
-　※開発中。仕様変更もありえます。
+　　今後もし % が実装できた場合、SiONにあわせて%ごとに独立にする可能性があります。
 ```
 
 ## 標準的なMML文法
