+++ lib/mml.js
@@ -591,21 +591,12 @@ function Sionic(mmldata) {
         return x | 0;
     }
 
-    function hexToInt8(x) {
-      return Int8Array.from([parseInt(x,16)])[0];
-    }
 
     var MMLCommands = [
         {
             re: /\#FM2OP@(\d+,\d+,\d+,\d+,\d+)/g,
             func: function(m) {
                 return { name: "#FM2OP@", val: m[1].split(",").map(toInt) };
-            }
-        },
-        {
-            re: /\#WAVB(\d+),([0-9A-F]{64})/g,
-            func: function(m) {
-                return { name: "#WAVB", val: [m[1], m[2].match(/.{2}/g).map(hexToInt8)] };
             }
         },
         {
@@ -630,12 +621,6 @@ function Sionic(mmldata) {
             re: /@(\d*)/g,
             func: function(m) {
                 return { name: "@", val: toInt(m[1]) };
-            }
-        },
-        {
-            re: /kt\-(\d*)/g,
-            func: function(m) {
-                return { name: "kt", val: - toInt(m[1]) };
             }
         },
         {
@@ -723,8 +708,9 @@ function Sionic(mmldata) {
         }
     ];
     
-    var SionAdapter = require("./sion-adapter");
-    var MMLCommands = SionAdapter.concat(MMLCommands);
+    var SionPreAdapter  = require("./sion-pre-adapter");
+    var SionPostAdapter = require("./sion-post-adapter");
+    var MMLCommands = SionPreAdapter.concat(MMLCommands).concat(SionPostAdapter);
 
     var sequencer = new MMLSequencer(mmldata);
