+++ lib/sion-post-adapter.js
@@ -1,13 +1,5 @@
 module.exports = [
     {
-        re: /\/\*[^\*]*\*\//g,
-        func: function(m) {
-            //C style Comments
-            //Do nothing
-            return { name: "comments" };
-        }
-    },
-    {
         re: /#[A-Z0-9 @]+{[^}]+}/g,
         func: function(m) {
             //SiON driver commands
@@ -16,6 +8,4 @@ module.exports = [
             return { name: "sion commands" };
         }
     }
-    
-    
 ]
